+++ eway-payment-gateway.php
@@ -3,16 +3,16 @@
 Plugin Name: Eway Payment Gateway
 Plugin URI: https://shop.webaware.com.au/downloads/eway-payment-gateway/
 Description: Take credit card payments via Eway in some popular WordPress plugins
-Version: 5.2.0
+Version: 5.2.1
 Author: WebAware
 Author URI: https://shop.webaware.com.au/
 Text Domain: eway-payment-gateway
 WC requires at least: 3.0
-WC tested up to: 8.3
+WC tested up to: 8.7
 */
 
 /*
-copyright (c) 2011-2023 WebAware Pty Ltd (email : support@webaware.com.au)
+copyright (c) 2011-2024 WebAware Pty Ltd (email : support@webaware.com.au)
 
 This program is free software; you can redistribute it and/or
 modify it under the terms of the GNU General Public License
@@ -38,7 +38,7 @@ define('EWAY_PAYMENTS_PLUGIN_FILE', __FILE__);
 define('EWAY_PAYMENTS_PLUGIN_ROOT', dirname(__FILE__) . '/');
 define('EWAY_PAYMENTS_PLUGIN_NAME', basename(dirname(__FILE__)) . '/' . basename(__FILE__));
 define('EWAY_PAYMENTS_MIN_PHP', '7.4');
-define('EWAY_PAYMENTS_VERSION', '5.2.0');
+define('EWAY_PAYMENTS_VERSION', '5.2.1');
 
 require EWAY_PAYMENTS_PLUGIN_ROOT . 'includes/functions-global.php';
 require EWAY_PAYMENTS_PLUGIN_ROOT . 'includes/class.Requires.php';